+++ fext/store/reducers/index.js
@@ -1,0 +1,14 @@
+/**
+ * fext root reducer
+ * @param {Object} state 
+ * @param {Object} action 
+ * @param {String} action.type action type
+ * @returns {Object} state
+ */
+function reducer(state = {}, action) {
+  switch (action.type) {
+    default: return state;
+  }
+}
+
+export default reducer;