+++ framework/app/components/data/index.js
@@ -1,4 +1,3 @@
 export { default as EdgeData } from './EdgeData';
 export { default as GraphData } from './GraphData';
-export { default as LayerData } from './LayerData';
 export { default as NodeData } from './NodeData';