+++ framework/app/store/reducers/metadata/index.js
@@ -3,6 +3,7 @@ import { SET_METADATA } from 'store/actionTypes';
 import graphsReducer from './graphs';
 import panelsReducer from './panels';
 import statusReducer from './status';
+import fextReducer from 'fext/store/reducers';
 
 /**
  * project metadata state reducer
@@ -19,12 +20,13 @@ function metadataReducer(state = {}, action) {
       return { ...state, ...data };
     };
     default: {
-      const { graphs, panels, status } = state;
+      const { graphs, panels, status, fext } = state;
       return {
         ...state,
         graphs: graphsReducer(graphs, action),
         panels: panelsReducer(panels, action),
-        status: statusReducer(status, action)
+        status: statusReducer(status, action),
+        fext: fextReducer(fext, action)
       };
     };
   }